Fix the playwright dependency version mismatch
This fixes #24.
It was disallowing us to run it with Firefox and Safari because
@playwright/test also adds the "playwright" from its dependencies to the
node_modules/.bin/ directory. That's was shadowing the playwright that
was added as the main dependency when `npx playwright install` was run
during the `postinstall` script execution. This was creating a
mismatch on the downloaded Firefox/Safari vs the Firefox/Safari the
`next` version wanted to run.